chore: Move create shul tests from form to view
